Changed the template method structure
