+++ src/DES.java
@@ -1,5 +1,18 @@
 package src;
 
-public class DES extends AES {
+import javax.crypto.SecretKey;
+import javax.crypto.spec.IvParameterSpec;
+
+public class DES extends Encryption {
+
+    public static String encrypt(String algorithm, String input, SecretKey key,
+            IvParameterSpec iv) {
+        return "";
+    }
+
+    public static String decrypt(String algorithm, String cipherText, SecretKey key,
+            IvParameterSpec iv) {
+        return "";
+    }
 
 }
